% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPOUMM.R
\name{analyseMCMCs}
\alias{analyseMCMCs}
\title{Extract data from an MCMC chain
This is an internal function.}
\usage{
analyseMCMCs(
  chains,
  stat = NULL,
  statName = "logpost",
  start,
  end,
  thinMCMC,
  as.dt = FALSE,
  k = NA,
  N = NA,
  ...
)
}
\arguments{
\item{chains, stat, statName, start, end, thinMCMC, as.dt, k, N, ...}{internal use.}
}
\description{
Extract data from an MCMC chain
This is an internal function.
}
