% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{chld}
\alias{chld}
\title{Node indices of the direct descendants of n in the phylogeny.}
\usage{
chld(tree, n)
}
\arguments{
\item{tree}{an object of class phylo}

\item{n}{an index of a node (root, internal or tip) in tree}
}
\value{
An integer vector.
}
\description{
Node indices of the direct descendants of n in the phylogeny.
}
