% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{coef.POUMM}
\alias{coef.POUMM}
\title{Extract maximum likelihood fitted parameters (coefficients) from a fitted 
POUMM  model.}
\usage{
\method{coef}{POUMM}(object, mapped = FALSE, ...)
}
\arguments{
\item{object}{An object of class POUMM.}

\item{mapped}{Logical indicating whether the standard POUMM parameters should
also be extracted.}

\item{...}{Not used; added for compatibility with generic function coef.}
}
\value{
A named vector with the fitted parameters of the model.
}
\description{
Extract maximum likelihood fitted parameters (coefficients) from a fitted 
POUMM  model.
}
\details{
The parameters extracted are the ones used as input to the model's
parMapping function.
}
