% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{covFunPOUMM}
\alias{covFunPOUMM}
\title{A vectorized expected covariance function for a given tree and a fitted POUMM
model}
\usage{
covFunPOUMM(object, corr = FALSE)
}
\arguments{
\item{object}{an S3 object of class POUMM}

\item{corr}{logical indicating if an expected correlation function 
should be returned
  For non-ultrametric trees, usually the mean root-tip distance is used.}
}
\value{
a function of three numerical parameters:
 tau - the phylogenetic distance between a two tips;
 tanc - the distance from the root to their most recent common ancestor.
 t - the root-tip distance (assuming that the two tips are at equal distance from the root)
}
\description{
A vectorized expected covariance function for a given tree and a fitted POUMM
model
}
