% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edgesFrom}
\alias{edgesFrom}
\title{Edge indices of the edges in tree starting from n}
\usage{
edgesFrom(tree, n)
}
\arguments{
\item{tree}{an object of class phylo}

\item{n}{an index of a node (root, internal or tip) in tree}
}
\value{
An integer vector.
}
\description{
Edge indices of the edges in tree starting from n
}
