% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{fitted.POUMM}
\alias{fitted.POUMM}
\title{Extract maximum likelihood expected genotypic values at the tips of a tree,
to which a POUMM model has been previously fitted}
\usage{
\method{fitted}{POUMM}(object, vCov = FALSE, ...)
}
\arguments{
\item{object}{An object of class POUMM.}

\item{vCov}{A logical indicating whether a list with the genotypic values and their variance covariance matrix should be returned or only a vector of the genotypic values (default is FALSE).}

\item{...}{Not used; added for compatibility with generic function fitted.}
}
\value{
If vCov == TRUE, a list with elements g - the genotypic values and
vCov - the variance-covariance matrix of these values for the specific tree,
observed values z and POUMM ML-fit. If vCov == FALSE, only the vector of 
genotypic values corresponding to the tip-labels in the tree is returned.
}
\description{
Extract maximum likelihood expected genotypic values at the tips of a tree,
to which a POUMM model has been previously fitted
}
