% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsPOUMM.R
\name{gPOUMM}
\alias{gPOUMM}
\title{Distribution of the genotypic values under a POUMM fit}
\usage{
gPOUMM(z, tree, g0, alpha, theta, sigma, sigmae)
}
\arguments{
\item{z}{A numeric vector of size length(tree$tip.label) representing the trait
values at the tip-nodes.}

\item{tree}{an object of class phylo}

\item{g0}{A numeric value at the root of the tree, genotypic contribution.}

\item{alpha, theta, sigma}{Numeric values, parameters of the OU model.}

\item{sigmae}{Numeric non-negative value (default 0). Specifies the standard
deviation of random environmental contribution (white noise) included in z.}
}
\value{
a list with elements V.g, V.g_1, mu.g, V.e, V.e_1, mu.e, V.g.poumm, mu.g.poumm.
}
\description{
Distribution of the genotypic values under a POUMM fit
}
