% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{logLik.POUMM}
\alias{logLik.POUMM}
\title{Extract maximum likelihood and degrees of freedom from a fitted POUMM model}
\usage{
\method{logLik}{POUMM}(object, ...)
}
\arguments{
\item{object}{An object of class POUMM.}

\item{...}{not used; included for compliance with generic function logLik.}
}
\description{
Extract maximum likelihood and degrees of freedom from a fitted POUMM model
}
