% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{nobs.POUMM}
\alias{nobs.POUMM}
\title{Number of tips in a phylogenetic tree, POUMM has been fit on.}
\usage{
\method{nobs}{POUMM}(object, ...)
}
\arguments{
\item{object}{An object of class POUMM.}

\item{...}{Not used; added for compatibility with generic function nobs.}
}
\value{
The number of tips in the tree, POUMM has been called on
}
\description{
Number of tips in a phylogenetic tree, POUMM has been fit on.
}
