% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nodeTimes}
\alias{nodeTimes}
\title{Calculate the time from the root to each node of the tree}
\usage{
nodeTimes(tree, tipsOnly = FALSE)
}
\arguments{
\item{tree}{An object of class phylo.}

\item{tipsOnly}{Logical indicating whether the returned results should be
truncated only to the tips of the tree.}
}
\value{
A vector of size the number of nodes in the tree (tips, root, 
  internal) containing the time from the root to the corresponding node in 
  the tree.
}
\description{
Calculate the time from the root to each node of the tree
}
