% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{residuals.POUMM}
\alias{residuals.POUMM}
\title{Extract maximum likelihood environmental contributions (residuals) at the tips of a tree, to which a POUMM model has been fitted.}
\usage{
\method{residuals}{POUMM}(object, ...)
}
\arguments{
\item{object}{An object of class POUMM.}

\item{...}{Not used; added for compatibility with generic function residuals.}
}
\value{
The vector of e-values (residuals) corresponding to the tip-labels in
 the tree.
}
\description{
Extract maximum likelihood environmental contributions (residuals) at the tips of a tree, to which a POUMM model has been fitted.
}
