% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsPOUMM.R
\name{rTrajectoryOUDef}
\alias{rTrajectoryOUDef}
\title{Generation of a random trajectory of an OU process starting from a given
initial state (only for test purpose)}
\usage{
rTrajectoryOUDef(z0, t, alpha, theta, sigma, steps = 1)
}
\arguments{
\item{z0}{Numeric value, initial state.}

\item{t}{Numeric value or vector of size steps, denoting the time-step(s).}

\item{alpha}{Numeric values, parameters of the OU process.}

\item{theta}{Numeric values, parameters of the OU process.}

\item{sigma}{Numeric values, parameters of the OU process.}

\item{steps}{Integer, number of steps.}
}
\description{
Generation of a random trajectory of an OU process starting from a given
initial state (only for test purpose)
}
\details{
Used for test purpose only. This is an internal function and is
  appropriate for small time-steps only.
}
