% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simulatePOUMMLikelihoodMainLoop}
\alias{simulatePOUMMLikelihoodMainLoop}
\title{Writes verbose messages of the order of tree traversal during likelihood calculation}
\usage{
simulatePOUMMLikelihoodMainLoop(tree)
}
\arguments{
\item{tree}{A phylo object.}
}
\value{
Nothing
}
\description{
Writes verbose messages of the order of tree traversal during likelihood calculation
}
