% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPOUMM.R
\name{statistics}
\alias{statistics}
\alias{statistics.POUMM}
\title{Extract statistics from sampled or inferred parameters of a 
POUMM fit}
\usage{
statistics(object)

\method{statistics}{POUMM}(object)
}
\arguments{
\item{object}{An object of class "POUMM".}
}
\description{
Extract statistics from sampled or inferred parameters of a 
POUMM fit
}
\details{
This is a generic method.
}
\section{Methods (by class)}{
\itemize{
\item \code{POUMM}: Relevant statistics from the sampled parameters of a
POUMM fit
}}

