% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specPOUMM.R
\name{validateSpecPOUMM}
\alias{validateSpecPOUMM}
\title{Validate a POUMM specification}
\usage{
validateSpecPOUMM(spec)
}
\arguments{
\item{spec}{A list object returned by one of the specifyPOUMM or specifyPMM
functions with possibly modified entries afterwards.}
}
\value{
The function either returns TRUE or exits with an error message if it
 finds a problem with the specificaiton.
}
\description{
Validate a POUMM specification
}
