% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteCachedResults.R
\docType{data}
\name{vignetteCachedResults}
\alias{vignetteCachedResults}
\title{Cached objects for the POUMM vignettes and examples}
\format{
This is a list containing the following named elements:
\describe{
  \item{g, z, e }{ numeric vectors of simulated genotypic values, phenotypic values and measurement errors. }
  \item{tree }{ a simulated phylogenetic tree. }
  \item{fitPOUMM, fitPOUMM2, fitH2tMean }{ POUMM fit objects to tree and z. }
}
}
\usage{
vignetteCachedResults
}
\description{
A list containing a simulated tree, trait-values and POUMM objects (model 
fits). To use these objects in examples you can load them into the global 
workspace with the command: 
`data(vignetteCachedResults); list2env(vignetteCachedResults, globalenv());`.
}
\keyword{datasets}
