% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{covFunPOUMM}
\alias{covFunPOUMM}
\title{A vectorized expected covariance function for a given tree and a fitted POUMM
model}
\usage{
covFunPOUMM(object, corr = FALSE)
}
\arguments{
\item{object}{an S3 object of class POUMM}

\item{corr}{logical indicating if an expected correlation function 
should be returned}
}
\value{
a function of a numerical parameter x denoting the phylogenetic distance
between a couple of tips.
}
\description{
A vectorized expected covariance function for a given tree and a fitted POUMM
model
}

